python/requirements.txt
amqp==1.4.9
anyjson==0.3.3
backports.shutil-get-terminal-size==1.0.0
billiard==3.3.0.23
celery==3.1.23
cycler==0.10.0
Cython==0.24
dask==0.10.0
decorator==4.0.10
enum34==1.1.6
futures==3.0.5
grpcio==0.15.0
h5py==2.6.0
ipython==4.2.1
ipython-genutils==0.1.0
kombu==3.0.35
leveldb==0.193
matplotlib==1.5.1
networkx==1.11
nose==1.3.7
numpy==1.11.1
pandas==0.18.1
pathlib2==2.1.0
pexpect==4.2.0
pickleshare==0.7.2
pika==0.10.0
Pillow==3.3.0
protobuf==3.0.0b3
ptyprocess==0.5.1
pyparsing==2.1.5
python-dateutil==1.5
python-gflags==3.0.5
pytz==2016.4
PyYAML==3.11
scikit-image==0.12.3
scipy==0.17.1
simplegeneric==0.8.1
six==1.10.0
toolz==0.8.0
traitlets==4.2.2
